Translated code for conditional breakpoints
Unlike Python (the language that I'm translating this from), in R, if I have a connection to a file and I read from it, then read from it *again* after appending to it elsewhere (e.g. through system() and the echo command), the connection will not recognize the appended line unless I use seek() to pull the connection back to the start of the file. This has consequences because it means we have to use some extra code in R (see readfromgbds()) to make sure we never read the same line more than once through gb.ds (otherwise, we'll get an infinite loop at a false conditional breakpoint, like so: checkdbgsink() -> dostep() -> checkdbgsink() -> dostep() -> etc).