Fixed bug in command parsing
The bug (which was present in the original debugR.py!) would only check the exact number of characters to identify a command (e.g. the first letter to check for "p"), which causes commands like "pcap" to never be triggered (since "pcap" is checked after "p"). This is no longer an issue now.