Successfully set up error sink for the screen terminal
This will allow reaction to error messages (e.g. to report the first line of the traceback to the user).

I did experiment with options(error=someFunc), where the appropriate response would be in someFunc(), but there are two issues: a) this option doesn't even pass the error message to someFunc(), b) if I do options(error=traceback), traceback() gets called before the "traceback is set" (so the traceback() call ends up reporting the previous traceback rather than the current one).

Also lost some time because I assumed "rw" was a valid file mode, and file() doesn't seem to confirm that the given file mode is valid.

I couldn't find a reason for this after a cursory search, but R doesn't seem to allow splitting the message connection, so if the errors are redirecting to a sink, the user won't be able to see them in the screen terminal (unless we write code to show the user the errors).