Helper func endscreen(). Ending this branch because it doesn't work, see commit description
Was trying to allow debugR to recover from errors (in debugR itself) while still displaying both the error message and the line number of the crash (from the traceback). But this is complicated because: 1) curses' endwin() has to clear the window (i.e. restore the contents of the old window), and 2) traceback() only works for uncaught errors. So this commit is more for recording a note.