(Finally) added assert() function allowing inspection of environment upon a false condition
This function has about the same effect as using browser(expr=(!condition)), and a more experienced user may prefer this. I said more about this in the comment preceding the definition of debugR::assert(). Since assert() is a wrapper, however, it allows me to print some helpful directions. This is helpful because one who is new to browser() may not know of ls(), and we aim to appeal to beginners here.

Also added example for trying out assert(). And the function is exported by the package.